+++ src/main/kotlin/fi/metatavu/pakkasmarja/services/erp/sap/BusinessPartnersController.kt
@@ -2,7 +2,6 @@ package fi.metatavu.pakkasmarja.services.erp.sap
 
 import com.fasterxml.jackson.databind.JsonNode
 import fi.metatavu.pakkasmarja.services.erp.sap.session.SapSessionController
-import fi.metatavu.pakkasmarja.services.erp.sap.utils.SapUtils
 import java.time.OffsetDateTime
 import javax.enterprise.context.ApplicationScoped
 import javax.inject.Inject
@@ -11,12 +10,9 @@ import javax.inject.Inject
  * The controller for business partners
  */
 @ApplicationScoped
-class BusinessPartnersController {
+class BusinessPartnersController: AbstractSapResourceController() {
     @Inject
     private lateinit var sapSessionController: SapSessionController
-
-    @Inject
-    private lateinit var sapUtils: SapUtils
 
     /**
      * Lists business partners
@@ -30,11 +26,11 @@ class BusinessPartnersController {
     fun listBusinessPartners(updatedAfter: OffsetDateTime?, firstResult: Int?, maxResults: Int?): ArrayList<JsonNode> {
         sapSessionController.createSapSession().use { sapSession ->
             val resourceUrl = "${sapSession.apiUrl}/BusinessPartners"
-            val updatedAfterFilter = updatedAfter?.let { "and ${sapUtils.createdUpdatedAfterFilter(it)}" } ?: ""
+            val updatedAfterFilter = updatedAfter?.let { "and ${createdUpdatedAfterFilter(it)}" } ?: ""
             val filter = "\$filter=(CardType eq 'cSupplier' $updatedAfterFilter)"
             val select = "\$select=CardCode,CardType,CardName,Phone1,Phone2,EmailAddress,BPAddresses,BPBankAccounts,FederalTaxID,VatLiable,UpdateDate,UpdateTime";
 
-            return sapUtils.getItemsAsJsonNodes(
+            return getItemsAsJsonNodes(
                 resourceUrl = resourceUrl,
                 filter = filter,
                 select = select,
