+++ src/main/kotlin/fi/metatavu/pakkasmarja/services/erp/sap/session/SapSession.kt
@@ -1,5 +1,6 @@
 package fi.metatavu.pakkasmarja.services.erp.sap.session
 
+import fi.metatavu.pakkasmarja.services.erp.sap.session.exception.SapSessionLogoutException
 import java.net.URI
 import java.net.http.HttpClient
 import java.net.http.HttpRequest
