+++ src/main/kotlin/fi/metatavu/pakkasmarja/services/erp/sap/session/exception/SapSessionLogoutException.kt
@@ -1,4 +1,4 @@
-package fi.metatavu.pakkasmarja.services.erp.sap.session
+package fi.metatavu.pakkasmarja.services.erp.sap.session.exception
 
 /**
  * Exception thrown when a SAP session logout fails
